Progress towards loading the FIBO AmigaOS sample.
* Turned off the [Obsolete] warnings in the M68kDisassembler as they hid other more serious warnings

* Fixed a bug in the m68k $nn(pc) address mode decoder.

* Support for more m68k effective address types in Rewriter.

* The address of the beginning very first CODE hunk is treated as an entry point.

+++ src/Core/Expressions/Constant.cs
@@ -261,7 +261,7 @@ namespace Decompiler.Core.Expressions
             return Convert.ToBoolean(GetValue());
         }
 
-        public ushort ToByte()
+        public byte ToByte()
         {
             return Convert.ToByte(GetValue());
         }
